# sstable: optimize twoLevelIterator, when exhausted, for SeekGE & SeekPrefixGE