This optimization is done when TrySeekUsingNext is set to true.
A new microbenchmark is added for this case and shows substantial
improvement. We expect this to be helpful when an sstable is not
part of a levelIter (since if the data is exhausted, the levelIter
will switch to a different sstable), or when only the bounds are
reached (even with a levelIter).

Note: two-level=true/exhausted=bounds does not show as much
improvement since the index block is loaded and a fast path will
be taken when twoLevelIterator calls into singleLevelIterator. If
the index block was not loaded, say because of block property
filter not matching, then the speedup should be similar to the
exhausted=file case.

```
name                                                                   old time/op    new time/op    delta
SeqSeekGEExhausted/two-level=false/exhausted=file/seek=ge-10             3.49ns ± 0%    3.46ns ± 0%   -0.72%  (p=0.008 n=5+5)
SeqSeekGEExhausted/two-level=false/exhausted=file/seek=prefix-ge-10      5.37ns ± 0%    5.34ns ± 1%   -0.67%  (p=0.032 n=5+5)
SeqSeekGEExhausted/two-level=false/exhausted=bounds/seek=ge-10           3.90ns ± 6%    3.76ns ± 0%   -3.72%  (p=0.008 n=5+5)
SeqSeekGEExhausted/two-level=false/exhausted=bounds/seek=prefix-ge-10    5.69ns ± 0%    5.61ns ± 0%   -1.40%  (p=0.008 n=5+5)
SeqSeekGEExhausted/two-level=true/exhausted=file/seek=ge-10               319ns ± 5%       4ns ± 0%  -98.72%  (p=0.008 n=5+5)
SeqSeekGEExhausted/two-level=true/exhausted=file/seek=prefix-ge-10        325ns ± 3%       5ns ± 3%  -98.53%  (p=0.008 n=5+5)
SeqSeekGEExhausted/two-level=true/exhausted=bounds/seek=ge-10            11.6ns ± 7%     4.1ns ± 0%  -64.67%  (p=0.016 n=5+4)
SeqSeekGEExhausted/two-level=true/exhausted=bounds/seek=prefix-ge-10     12.4ns ±12%     4.7ns ± 0%  -61.89%  (p=0.008 n=5+5)
```